[C-API/Service] Refactoring the service API for maintainability and readability
To support various database systems, this patch defines the interface
class IMLServiceDB and implements MLServiceLevelDB using LevelDB. If the
backend DB is changed, we just add a new class for the new DB without
changing the existing code. That increases the maintainability.
Moreover, this patch updates the signature of user-level APIs to
increase readability.

Signed-off-by: Sangjung Woo <sangjung.woo@samsung.com>