[C-API/Test] Add Mock testcase for Service API
ML Service API uses an external database library and it could raise
unexpected errors such as IO Exception. To handle their exceptions,
lots of error handling code was added but it is hard to test because
exceptions are not raised in normal cases. In this patch, we use GMock
for dependency injection of external libraries. We can define the
expected exception when calling an external API so can verify its operation.

Signed-off-by: Sangjung Woo <sangjung.woo@samsung.com>