[Refactoring] do refactoring the NNFW unit test to improve maintainability
This patch applies the Google Test Fixture to the unit test of NNFW
Inference MLAPI test. It removes the duplicate codes such as finalizing
the handle of TensorsInfo, TensorsData, Single, and Pipeline in each
test case. As a result, it helps developers focus on the test case
scenario without considering trivial settings such as model loading,
or memory allocation.

Signed-off-by: Sangjung Woo <sangjung.woo@samsung.com>