[Service/PkgMgr] Add skeleton code for the package manager
To get the event from the package manager, the callback function should
be registered in ML Agent. This patch adds the skeleton code to get the
install, uninstall, and upgrade event of the resource package from the
package manager. Related work such as update the database of ML Agent
should be needed.

Signed-off-by: Sangjung Woo <sangjung.woo@samsung.com>